cambia el tipo de dato del método setPrecioXpersona de int a double para mejorar la precisión del precio

+++ app/src/main/java/es/unizar/eina/M12_camping/database/Parcela.java
@@ -69,7 +69,7 @@ public class Parcela {
     }
 
     /** Permite actualizar el precio por persona de una parcela */
-    public void setPrecioXpersona(int precioXpersona) {
+    public void setPrecioXpersona(double precioXpersona) {
         this.precioXpersona = precioXpersona;
     }
